+++ private/bufpkg/bufimage/bufimageutil/bufimageutil.go
@@ -450,17 +450,15 @@ func (t *transitiveClosure) addElement(
 
 	switch typedDescriptor := descriptor.(type) {
 	case *descriptorpb.FileDescriptorProto:
-		typeNames, ok := imageIndex.FileTypes[typedDescriptor.GetName()]
 		// If the file we are attempting to include is empty (has no types defined), it is still
 		// valid for inclusion.
-		if ok {
-			// A file includes all elements. The types are resolved in the image index
-			// to ensure all nested types are included.
-			for _, typeName := range typeNames {
-				typeInfo := imageIndex.ByName[typeName]
-				if err := t.addElement(typeInfo.element, "", false, imageIndex, opts); err != nil {
-					return err
-				}
+		typeNames := imageIndex.FileTypes[typedDescriptor.GetName()]
+		// A file includes all elements. The types are resolved in the image index
+		// to ensure all nested types are included.
+		for _, typeName := range typeNames {
+			typeInfo := imageIndex.ByName[typeName]
+			if err := t.addElement(typeInfo.element, "", false, imageIndex, opts); err != nil {
+				return err
 			}
 		}