+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix type filtering on `buf generate` for empty files, files with no declared types.
 - Fix CEL check on `buf lint` for predefined `rules` variables.
 - Fix `buf config migrate` to filter out removed rules. 
+- Add ppc64le binaries for Linux to releases.
 
 ## [v1.53.0] - 2025-04-21