+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [v1.54.0] - 2025-05-12
 
 - Add `CSR` category to breaking rules.
 - Add support for local bufplugins for `protoc-gen-buf-breaking` and `protoc-gen-buf-lint`.
@@ -1371,7 +1371,7 @@ buf check breaking proto --against .git#branch=master,subdir=proto
 
 Initial beta release.
 
-[Unreleased]: https://github.com/bufbuild/buf/compare/v1.53.0...HEAD
+[v1.54.0]: https://github.com/bufbuild/buf/compare/v1.53.0...v1.54.0
 [v1.53.0]: https://github.com/bufbuild/buf/compare/v1.52.1...v1.53.0
 [v1.52.1]: https://github.com/bufbuild/buf/compare/v1.52.0...v1.52.1
 [v1.52.0]: https://github.com/bufbuild/buf/compare/v1.51.0...v1.52.0